fix signedness warnings

+++ geometry/estimate-point-normals.cpp
@@ -131,7 +131,7 @@ void reorientNormals(const std::vector<math::Point3>& pc
 {
     std::vector<BvhDisk> disks(pc.size());
     UTILITY_OMP(parallel for shared(pc))
-    for (int64_t i = 0; i < pc.size(); ++i) {
+    for (int64_t i = 0; i < (int64_t)pc.size(); ++i) {
         disks[i] = BvhDisk(pc[i], normals[i], pointRadius, i);
     }
 
@@ -151,7 +151,7 @@ void reorientNormals(const std::vector<math::Point3>& pc
     // assuming the topmost points (roofs) have z>0 orientation.
     LOG(info2) << "Estimating normal orientations";
     UTILITY_OMP(parallel for shared(orderZ))
-    for (int64_t rankZ = 0; rankZ < orderZ.size(); ++rankZ) {
+    for (int64_t rankZ = 0; rankZ < (int64_t)orderZ.size(); ++rankZ) {
         const std::uint32_t i = orderZ[rankZ];
         int doFlip = 0;
         for (math::Point3 dir : ALL_DIRS) {
@@ -184,7 +184,7 @@ void reorientNormals(const std::vector<math::Point3>& pc
     std::vector<uint8_t> doFlip(pc.size(), false);
     std::vector<math::Points3::const_iterator> neighs;
     UTILITY_OMP(parallel for private(neighs))
-    for (int64_t i = 0; i < pc.size(); ++i) {
+    for (int64_t i = 0; i < (int64_t)pc.size(); ++i) {
         neighs.clear();
         tree.range(pc[i], 2 * pointRadius, neighs);
 
@@ -198,7 +198,7 @@ void reorientNormals(const std::vector<math::Point3>& pc
         }
     }
     UTILITY_OMP(parallel for shared(pc))
-    for (int64_t i = 0; i < pc.size(); ++i) {
+    for (int64_t i = 0; i < (int64_t)pc.size(); ++i) {
         if (doFlip[i]) {
             normals[i] *= -1;
         }
